backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
pillow
python-decouple
dj-database-url
psycopg2-binary
whitenoise
gunicorn
